# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Digital Tolk Translation API",
        "description": "OpenAPI documentation for the versioned Laravel API.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/"
        }
    ],
    "paths": {
        "/api/v1/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register a new user and issue a Sanctum token",
                "operationId": "registerUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe",
                                        "maxLength": 255
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com",
                                        "maxLength": 255
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "Password123!",
                                        "minLength": 8
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "Password123!",
                                        "minLength": 8
                                    }
                                },
                                "type": "object"
                            },
                            "example": {
                                "name": "John Doe",
                                "email": "john@example.com",
                                "password": "Password123!",
                                "password_confirmation": "Password123!"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Registered successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                },
                                "example": {
                                    "success": true,
                                    "message": "Registered",
                                    "data": {
                                        "user": {
                                            "id": 1,
                                            "name": "John Doe",
                                            "email": "john@example.com",
                                            "email_verified_at": null,
                                            "created_at": "2026-06-24T10:00:00Z",
                                            "updated_at": "2026-06-24T10:00:00Z"
                                        },
                                        "token": "1|plain-text-sanctum-token",
                                        "token_type": "Bearer"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Authenticate a user and issue a Sanctum token",
                "operationId": "loginUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "Password123!"
                                    }
                                },
                                "type": "object"
                            },
                            "example": {
                                "email": "john@example.com",
                                "password": "Password123!"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Logged in successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                },
                                "example": {
                                    "success": true,
                                    "message": "Logged in",
                                    "data": {
                                        "user": {
                                            "id": 1,
                                            "name": "John Doe",
                                            "email": "john@example.com",
                                            "email_verified_at": null,
                                            "created_at": "2026-06-24T10:00:00Z",
                                            "updated_at": "2026-06-24T10:00:00Z"
                                        },
                                        "token": "1|plain-text-sanctum-token",
                                        "token_type": "Bearer"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "example": {
                                    "success": false,
                                    "message": "Invalid credentials",
                                    "errors": null
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Revoke the current Sanctum token",
                "operationId": "logoutUser",
                "responses": {
                    "200": {
                        "description": "Logged out successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                },
                                "example": {
                                    "success": true,
                                    "message": "Logged out",
                                    "data": null
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "example": {
                                    "success": false,
                                    "message": "Unauthenticated.",
                                    "errors": null
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/me": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Get the authenticated user",
                "operationId": "meUser",
                "responses": {
                    "200": {
                        "description": "Authenticated user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                },
                                "example": {
                                    "success": true,
                                    "message": "Success",
                                    "data": {
                                        "id": 1,
                                        "name": "John Doe",
                                        "email": "john@example.com",
                                        "email_verified_at": null,
                                        "created_at": "2026-06-24T10:00:00Z",
                                        "updated_at": "2026-06-24T10:00:00Z"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "example": {
                                    "success": false,
                                    "message": "Unauthenticated.",
                                    "errors": null
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "SuccessResponse": {
                "required": [
                    "success",
                    "message"
                ],
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Success"
                    },
                    "data": {
                        "type": "object",
                        "example": {
                            "id": 1
                        },
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "required": [
                    "success",
                    "message"
                ],
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Error"
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "message": [
                                "Something went wrong."
                            ]
                        },
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ValidationErrorResponse": {
                "required": [
                    "success",
                    "message",
                    "errors"
                ],
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "email": [
                                "The email field is required."
                            ],
                            "password": [
                                "The password field is required."
                            ]
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "Token"
            }
        }
    },
...
}